feat: - Created a project with two submodules(webapp and provider) - Implemented RestTemplate to make call to Spring Boot provider
