--- README.md
@@ -1,2 +1,0 @@
-# schedule-now
-ScheduleNow will be a simple scheduling application for freelancers or contractors that need to organize their time and customers in one easy space.
