+++ ScheduleNowProvider/src/main/java/com/bruno/provider/ScheduleNowApplication.java
@@ -1,0 +1,16 @@
+package com.bruno.provider;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EnableJpaRepositories
+@EntityScan
+public class ScheduleNowApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ScheduleNowApplication.class, args);	
+	}
+}
