+++ ScheduleNowProvider/src/main/java/com/bruno/provider/model/User.java
@@ -1,0 +1,11 @@
+package com.bruno.provider.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class User {
+	private String username;
+	private String password;
+}
