+++ ScheduleNowProvider/src/main/java/com/bruno/provider/repository/JobDao.java
@@ -1,0 +1,9 @@
+package com.bruno.provider.repository;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.bruno.provider.model.Job;
+
+public interface JobDao extends JpaRepository<Job, Long> {
+}
