+++ ScheduleNowProvider/src/main/java/com/bruno/provider/service/JobService.java
@@ -1,0 +1,18 @@
+package com.bruno.provider.service;
+
+import java.util.List;
+
+import com.bruno.provider.model.Job;
+
+public interface JobService {
+
+	void saveJob(Job job);
+	
+	List<Job> getAllJobs();
+	
+	Job getById(Long id);
+
+	Job update(Long id, Job job);
+
+	void deleteJob(Long id);
+}
