+++ ScheduleNowProvider/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+server.port = 8081
+spring.h2.console.enabled = true
+
+spring.datasource.url = jdbc:h2:mem:test;DB_CLOSE_DELAY=-1
+spring.datasource.driverClassName = org.h2.Driver
+spring.datasource.username = sa
+spring.datasource.password = sa
+
+spring.jpa.defer-datasource-initialization=true
+spring.jpa.database-platform = org.hibernate.dialect.H2Dialect
+logging.level.org.hibernate.SQL = DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder = TRACE
+hibernate.hbm2ddl.auto = create-drop
