+++ ScheduleNowProvider/src/main/resources/data.sql
@@ -1,0 +1,1 @@
+INSERT INTO JOB (ID, ADDRESS, CUSTOMER_NAME, DATE_TIME, DESCRIPTION, EMAIL, PHONE) VALUES (1, '12 Main St', 'Bruno', {ts '2022-05-21 12:00:00.00'}, 'Yard Work', 'brunoara12@gmail.com', '973-222-3122'), (2,'31 Broadway Ave', 'Alex Leo',   {ts '2022-05-21 14:05:00.00'},'Tree Cutting', 'a.leo@gmail','201-959-4522'), (3, '421 Madison Rd', 'Ariel Lazzu', {ts '2022-04-19 12:35:00.00'}, 'Pool Work', '','201-654-8941');
