+++ pom.xml
@@ -1,100 +1,14 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.bruno</groupId>
-  <artifactId>ScheduleNow</artifactId>
-  <packaging>war</packaging>
+  <artifactId>schedule-now</artifactId>
   <version>0.0.1-SNAPSHOT</version>
-  <name>ScheduleNow Maven Webapp</name>
-  <url>http://maven.apache.org</url>
-  <properties>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-</properties>
-  <dependencies>
+  <packaging>pom</packaging>
+  <name>ScheduleNow</name>
+  <description>Root</description>
   
-  <!-- https://mvnrepository.com/artifact/javax.servlet/javax.servlet-api -->
-	<dependency>
-	    <groupId>javax.servlet</groupId>
-	    <artifactId>javax.servlet-api</artifactId>
-	    <version>3.1.0</version>
-	    <scope>provided</scope>
-	</dependency>
-	
-  	<!-- https://mvnrepository.com/artifact/org.springframework/spring-webmvc -->
-	<dependency>
-	    <groupId>org.springframework</groupId>
-	    <artifactId>spring-webmvc</artifactId>
-	    <version>5.1.8.RELEASE</version>
-	</dependency>
-
-	<dependency>
-	    <groupId>org.springframework</groupId>
-	    <artifactId>spring-orm</artifactId>
-	    <version>5.1.8.RELEASE</version>
-	</dependency>
-
-  	<!-- https://mvnrepository.com/artifact/javax.servlet.jsp.jstl/jstl-api -->
-	<dependency>
-	    <groupId>javax.servlet</groupId>
-	    <artifactId>jstl</artifactId>
-	    <version>1.2</version>
-	</dependency>
- 
-	<dependency>                             
-		<groupId>org.slf4j</groupId>         
-		<artifactId>slf4j-simple</artifactId>
-	    <version>1.7.30</version>  
-	</dependency>
-	
-	<dependency>
-		<groupId>com.fasterxml.jackson.core</groupId>
-		<artifactId>jackson-databind</artifactId>
-		<version>2.9.4</version>
-	</dependency>
-
-	<dependency>
-        <groupId>taglibs</groupId>
-        <artifactId>standard</artifactId>
-        <version>1.1.2</version>
-    </dependency>
-
-
-	<!-- https://mvnrepository.com/artifact/org.hibernate/hibernate-core -->
-	<dependency>
-	    <groupId>org.hibernate</groupId>
-	    <artifactId>hibernate-core</artifactId>
-	    <version>5.4.33.Final</version>
-	</dependency>
-
-	<dependency>
-	    <groupId>com.h2database</groupId> 
-	    <artifactId>h2</artifactId>
-	    <version>1.4.197</version>
-	    <scope>runtime</scope>
-	</dependency>
-
-
-
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-  
-  <build>
-  	<plugins>
-	  	<plugin>
-		    <groupId>org.apache.maven.plugins</groupId>
-		    <artifactId>maven-war-plugin</artifactId>
-		    <version>3.3.1</version>
-		</plugin>
-	</plugins>
-    <finalName>ScheduleNow</finalName>
-  </build>
-  
-  
+  <modules>
+  	<module>webapp</module>
+  	<module>ScheduleNowProvider</module>
+  </modules>
 </project>
-
