--- src/main/java/com/bruno/config/WebInitializer.java
@@ -1,22 +1,0 @@
-package com.bruno.config;
-
-import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
-
-public class WebInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
-
-	@Override
-	protected Class<?>[] getRootConfigClasses() {
-		return new Class[] {ApplicationMvcConfig.class};
-	}
-
-	@Override
-	protected Class<?>[] getServletConfigClasses() {
-		return new Class[] {};
-	}
-
-	@Override
-	protected String[] getServletMappings() {
-		return new String[] {"/"};
-	}
-
-}
