--- src/main/java/com/bruno/dto/JobDto.java
@@ -1,50 +1,0 @@
-package com.bruno.dto;
-
-public class JobDto {
-
-	private String customerName;
-	private String address;
-	private String email;
-	private String phone;
-	private Long date;
-	private String description;
-	
-	public String getCustomerName() {
-		return customerName;
-	}
-	public void setCustomerName(String customerName) {
-		this.customerName = customerName;
-	}
-	public String getAddress() {
-		return address;
-	}
-	public void setAddress(String address) {
-		this.address = address;
-	}
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	public String getPhone() {
-		return phone;
-	}
-	public void setPhone(String phone) {
-		this.phone = phone;
-	}
-	public Long getDate() {
-		return date;
-	}
-	public void setDate(Long date) {
-		this.date = date;
-	}
-	public String getDescription() {
-		return description;
-	}
-	public void setDescription(String description) {
-		this.description = description;
-	}
-	
-	
-}
