--- src/main/java/com/bruno/models/User.java
@@ -1,21 +1,0 @@
-package com.bruno.models;
-
-public class User {
-
-	private String username;
-	private String password;
-	
-	public String getUsername() {
-		return username;
-	}
-	public void setUsername(String username) {
-		this.username = username;
-	}
-	public String getPassword() {
-		return password;
-	}
-	public void setPassword(String password) {
-		this.password = password;
-	}
-	
-}
