--- src/main/resources/application.properties
@@ -1,9 +1,0 @@
-jdbc.driverClassName = org.h2.Driver
-jdbc.url = jdbc:h2:tcp://localhost/~/test
-jdbc.username = sa
-jdbc.password = sa
-
-hibernate.dialect = org.hibernate.dialect.H2Dialect
-hibernate.show_sql = true
-hibernate.format_sql = true
-hibernate.hbm2ddl.auto = create-drop
