+++ webapp/src/main/java/com/bruno/webapp/model/Job.java
@@ -1,0 +1,80 @@
+package com.bruno.webapp.model;
+
+import java.util.Date;
+
+public class Job {
+
+	private Long id;
+	private String customerName;
+	private String address;
+	private String email;
+	private String phone;
+	private Date dateTime;
+	private String description;
+	
+	
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public String getCustomerName() {
+		return customerName;
+	}
+
+	public void setCustomerName(String customerName) {
+		this.customerName = customerName;
+	}
+
+	public String getAddress() {
+		return address;
+	}
+
+	public void setAddress(String address) {
+		this.address = address;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	public String getPhone() {
+		return phone;
+	}
+
+	public void setPhone(String phone) {
+		this.phone = phone;
+	}
+
+	public Date getDateTime() {
+		return dateTime;
+	}
+
+	public void setDateTime(Date dateTime) {
+		this.dateTime = dateTime;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String descriptionMdsamdsaim) {
+		this.description = descriptionMdsamdsaim;
+	}
+
+	@Override
+	public String toString() {
+		return "Job [id=" + id + ", customerName=" + customerName + ", address=" + address + ", email=" + email
+				+ ", phone=" + phone + ", dateTime=" + dateTime + ", description=" + description + "]";
+	}
+
+
+	
+}
