+++ webapp/src/main/java/com/bruno/webapp/model/User.java
@@ -1,0 +1,21 @@
+package com.bruno.webapp.model;
+
+public class User {
+
+	private String username;
+	private String password;
+	
+	public String getUsername() {
+		return username;
+	}
+	public void setUsername(String username) {
+		this.username = username;
+	}
+	public String getPassword() {
+		return password;
+	}
+	public void setPassword(String password) {
+		this.password = password;
+	}
+	
+}
