+++ webapp/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+service.url=http://localhost:8081
