+++ webapp/src/main/webapp/static/js/services/scheduling_service.js
